+++ src/core/film.go
@@ -5,8 +5,8 @@ type Film interface {
 	Splat(sample *Sample, L *Spectrum)
 	GetSampleExtent() (xstart, xend, ystart, yend int)
 	GetPixelExtent() (xstart, xend, ystart, yend int)
-	UpdateDisplay(x0, y0, x1, y1 int, splatScale float32)
-	WriteImage(splatScale float32)
+	UpdateDisplay(x0, y0, x1, y1 int, splatScale float64)
+	WriteImage(splatScale float64)
 	XResolution() int
 	YResolution() int
 }
